Add support for UndefinedBehaviorSanitizer
Changes the `SLANG_ENABLE_ASAN` CMake option to also enable
UndefinedBehaviorSanitizer (UBSan) if the C++ compiler is Clang or GCC
(MSVC only supports ASan, not UBSan).

This PR only does the changes necessary for building Slang with UBSan
and running `slang-test` without it crashing.

UBSan does not abort execution after reporting an error (unlike ASan),
and as there are many UBSan errors to fix, they will be fixed in other
PRs.

However, one of UBSan's type checks was causing a segfault/crash as it
was dereferencing a null `this` pointer. The root cause is that
`getCurrentASTBuilder()` was returning a null pointer in some test
cases. This PR changes global session initialization to set the
session's root AST builder as current AST builder. The previous AST
builder is restored when the session is destroyed. This also fixes some
of the UBSan errors (member function calls on null `ASTBuilder`
pointers).

Fixes #9099.